+++ internal/install/command.go
@@ -9,6 +9,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/newrelic/newrelic-cli/internal/client"
+	"github.com/newrelic/newrelic-cli/internal/config"
 	"github.com/newrelic/newrelic-cli/internal/credentials"
 	"github.com/newrelic/newrelic-client-go/newrelic"
 )
@@ -21,6 +22,8 @@ var (
 	skipIntegrations   bool
 	skipLoggingInstall bool
 	testMode           bool
+	debug              bool
+	trace              bool
 )
 
 // Command represents the install command.
@@ -39,6 +42,12 @@ var Command = &cobra.Command{
 		}
 
 		client.WithClientAndProfile(func(nrClient *newrelic.NewRelic, profile *credentials.Profile) {
+			if trace {
+				log.SetLevel(log.TraceLevel)
+			} else if debug {
+				log.SetLevel(log.DebugLevel)
+			}
+
 			err := assertProfileIsValid(profile)
 			if err != nil {
 				log.Fatal(err)
@@ -48,7 +57,7 @@ var Command = &cobra.Command{
 
 			// Run the install.
 			if err := i.Install(); err != nil {
-				log.Fatalf("Could not install New Relic: %s", err)
+				log.Fatalf("Could not install New Relic: %s, check the install log for details: %s", err, config.DefaultLogFile)
 			}
 		})
 	},
@@ -72,4 +81,6 @@ func init() {
 	Command.Flags().BoolVarP(&skipIntegrations, "skipIntegrations", "r", false, "skips installation of recommended New Relic integrations")
 	Command.Flags().BoolVarP(&skipLoggingInstall, "skipLoggingInstall", "l", false, "skips installation of New Relic Logging")
 	Command.Flags().BoolVarP(&testMode, "testMode", "t", false, "fakes operations for UX testing")
+	Command.Flags().BoolVar(&debug, "debug", false, "debug level logging")
+	Command.Flags().BoolVar(&trace, "trace", false, "trace level logging")
 }