+++ internal/install/types/recipe.go
@@ -1,12 +1,10 @@
 package types
 
 type Recipe struct {
-	ID   string `json:"id"`
-	File string `json:"file"`
-	Name string `json:"name"`
-	// TODO: sort out DisplayName vs Name
-	// nolint:govet
-	DisplayName    string     `json:"name"`
+	ID             string     `json:"id"`
+	File           string     `json:"file"`
+	Name           string     `json:"name"`
+	DisplayName    string     `json:"displayName"`
 	Description    string     `json:"description"`
 	Repository     string     `json:"repository"`
 	Keywords       []string   `json:"keywords"`
@@ -29,6 +27,8 @@ type LogMatch struct {
 type LogMatchAttributes struct {
 	LogType string `yaml:"logtype"`
 }
+
+type RecipeVars map[string]string
 
 // AddVar is responsible for including a new variable on the recipe Vars
 // struct, which is used by go-task executor.